chore: 쿼리 생성 함수 파일 분리
